flask-code/requirements.txt

Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
pysqlite3
flask-cors